ref(spans): Reorganize the segment consumer code
